+++ src/demo/Demo.tsx
@@ -3,11 +3,12 @@ import React from "react";
 import styles from "./Demo.module.scss";
 import { Menu } from "./Menu";
 import { Main } from "./Main";
+import { componentsMap } from "./componentsMap";
 
 const Demo: React.FC = (props) => {
     return (
         <div className={styles.container}>
-            <Menu />
+            <Menu list={componentsMap} />
             <Main />
         </div>
     );