+++ src/demo/componentsMap.ts
@@ -4,13 +4,43 @@ import { HeaderDemo } from "./components/layout/header/Header";
 import { StickyHeaderDemo } from "./components/layout/header/StickyHeader";
 import { ListItemDemo } from "./components/layout/list/List.Item";
 
-const componentsMap: { [key: string]: (typeof React.Component) | React.FC } = {
-    checkbox: CheckboxDemo,
-    header: HeaderDemo,
-    stickyHeader: StickyHeaderDemo,
-    listItem: ListItemDemo,
+type AnyComponent = React.ComponentClass | React.FC;
+
+interface TheMap {
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+    [key: string]: TheMapItem;
+}
+
+interface TheMapItem {
+    name: string;
+    Component: AnyComponent;
+    children?: TheMap;
+}
+
+const componentsMap: TheMap = {
+    Checkbox: {
+        name: "Checkbox",
+        Component: CheckboxDemo,
+    },
+    Header: {
+        name: "Header",
+        Component: HeaderDemo,
+    },
+    StickyHeader: {
+        name: "Sticky header",
+        Component: StickyHeaderDemo,
+    },
+    ListItem: {
+        name: "List & Item",
+        Component: ListItemDemo,
+    },
 };
 
 export {
     componentsMap,
+};
+
+export type {
+    TheMap,
+    TheMapItem,
 };