+++ src/components/ui/directionPad/Button.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import styles from "./Pad.module.scss";
+
+interface Props {
+    onClick?: () => void;
+}
+
+const PadButton: React.FC<Props> = (props) => {
+    return (
+        <button {...props} className={styles.pad__button}><span className={styles.pad__dot} /></button>
+    );
+};
+
+export { PadButton };