+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Header + StickyHeader component
 - List + Item components
+- multi-level menu support in demo
 ### Fixed
 - scrolling on the demo page